+++ packages/search/global-search/src/components/confluence/ConfluenceSearchResults.tsx
@@ -14,7 +14,11 @@ import {
   getConfluenceAdvancedSearchLink,
 } from '../SearchResultsUtil';
 import AnalyticsEventFiredOnMount from '../analytics/AnalyticsEventFiredOnMount';
-import { buildScreenEvent, Screen } from '../../util/analytics';
+import { buildScreenEvent, Screen } from '../../util/analytics-util';
+
+export const MAX_PAGES_BLOGS_ATTACHMENTS = 8;
+export const MAX_SPACES = 3;
+export const MAX_PEOPLE = 3;
 
 let preQueryScreenCounter = 0;
 let postQueryScreenCounter = 0;
@@ -128,12 +132,13 @@ const renderSearchResults = (
   return [
     renderObjectsGroup(
       'Pages, blogs and attachments',
-      take(objectResults, 8),
+      take(objectResults, MAX_PAGES_BLOGS_ATTACHMENTS),
       query,
     ),
-    renderSpacesGroup('Spaces', take(spaceResults, 3), query),
-    renderPeopleGroup('People', take(peopleResults, 3), query),
+    renderSpacesGroup('Spaces', take(spaceResults, MAX_SPACES), query),
+    renderPeopleGroup('People', take(peopleResults, MAX_PEOPLE), query),
     renderAdvancedSearchGroup(query),
+
     <AnalyticsEventFiredOnMount
       key="postQueryScreenEvent"
       onEventFired={() => postQueryScreenCounter++}