+++ packages/editor/editor-wikimarkup-transformer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-wikimarkup-transformer
 
+## 1.1.9
+- [patch] Updated dependencies [af0cde6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/af0cde6)
+  - @atlaskit/editor-core@74.0.0
+  - @atlaskit/editor-test-helpers@4.2.2
+  - @atlaskit/editor-json-transformer@3.1.7
+
 ## 1.1.8
 - [patch] Add missing dependencies to packages to get the website to build [99446e3](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/99446e3)