+++ packages/media/media-filmstrip/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-filmstrip
 
+## 9.0.7
+- [patch] Updated dependencies [42ee1ea](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/42ee1ea)
+  - @atlaskit/media-test-helpers@14.0.6
+  - @atlaskit/media-core@21.0.0
+  - @atlaskit/media-card@29.1.8
+
 ## 9.0.6
 - [patch] Add missing dependencies to packages to get the website to build [99446e3](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/99446e3)